Add files via upload
Fixed bug, the YTMusic menu fast search was saving urls as www. and slow search did not have the www. in the url. This would not have been an issue unless your looking for duplicate urls to delete. So going forward this issue has been fixed. You will have to manualy delete all the old duplicates.