Add files via upload
Added less to the list of dependencies for Linux, in the dependencies.txt
Linux v2.52.5_test | macOS v2.52.5_test | Termux v2.52.5_test
Changed the format of the code in yt-playlist, install.sh and uninstall.sh
Fixed the Playlist menu option number 4.
Changed the error message on the function that checks for YTMusic Radio urls.
Fixed the function that check for YTMusic Radio urls.
Fixed bug, YouTube Music Radio is not support. Added an error message.